# Conflicts:
#	yashua-portal/src/main/resources/statics/css/css.css

+++ modules/account/account-api/src/main/java/com/lebaoxun/modules/account/service/IUserMessageService.java
@@ -64,7 +64,7 @@ public interface IUserMessageService {
      */
     @RequestMapping("/account/usermessage/delete")
     ResponseMessage delete(@RequestParam("adminId") Long adminId,
-    		@RequestBody Integer[] ids);
+    		@RequestBody Long[] ids);
     
 }
